flux2imd - 2025.1.30.1: migraged to common utlilties

+++ flux2imd/sectorManager.c
@@ -75,7 +75,7 @@ static unsigned slotAt(int pos, bool isIdam) {
     if (pos <= 0)
         return -pos;
 
-    if (prevSlot < 0 || pos + POSJITTER < (isIdam ? prevIdamPos : prevDataPos)) {       // initialise either explicit or due to seek from start
+    if (prevSlot < 0 || pos + POSJITTER < (int)(isIdam ? prevIdamPos : prevDataPos)) {       // initialise either explicit or due to seek from start
         curSpacing = curFormat->spacing;
         if (getHsCnt())
             minSpacing = maxSpacing = curSpacing;
